requirements.txt
certifi==2017.7.27.1
click==6.7
Flask==0.12.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-telegram-bot==7.0.1
Werkzeug==0.12.2
gunicorn==19.7.1
